+++ app/(afterLogin)/detailPost/[postId]/_components/comment/createReComment.tsx
@@ -1,11 +1,11 @@
 "use client"
 
-import { useCreateReComment } from "@/hook/useCommentMutation"
+import { useCreateReComment } from "@/libs/reactQuery/useCommentMutation"
 import { CommentProps } from "./type"
 import { useState } from "react"
 import SuccessToFailModal from "@/components/commons/successToFailModal"
-import { putCommentRequest } from "@/hook/type"
 import Button from "@/components/commons/button"
+import { putCommentRequest } from "@/libs/reactQuery/type"
 
 const CreateReComment = ({ postId, commentId, refetch, onReplySuccess }: Partial<CommentProps>) => {
     const [content, setContent] = useState<string>("")
