+++ hook/useUpdatePost.ts
@@ -4,7 +4,7 @@ import { FormEvent, useState } from "react"
 import { useCreatePostStore, useUpdatePostDataStore } from "@/libs/store"
 import { postPost } from "@/apis/postApi"
 import { memos, UpdatePost } from "@/types/post"
-import { useUpdateFreePost } from "./usePostMutation"
+import { useUpdateFreePost } from "../libs/reactQuery/usePostMutation"
 import { useCommonUpdatePost, useInitializeFormData } from "./updatePostFunctions"
 import dayjs from "dayjs"
 import utc from "dayjs/plugin/utc"
