+++ libs/reactQuery/usePostMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient, UseMutationResult } from "@tanstack/react-query"
 import { deletePost, getPost, putFreePost } from "@/apis/postApi"
-import { updateFreeProps, useGetPostProps } from "./type"
 import { Post, UpdatePost } from "@/types/post"
+import { updateFreeProps, useGetPostProps } from "./type"
 
 export const useGetPost = (): UseMutationResult<Post[], Error, useGetPostProps> => {
     const queryClient = useQueryClient()
